% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localComm.R
\docType{class}
\name{localComm-class}
\alias{localComm-class}
\alias{localComm}
\title{Local community}
\usage{
localComm(indSpecies, indTrait, indSeqs)
}
\description{
An S4 class to specify the state of a local community
}
\section{Slots}{

\describe{
\item{\code{indSpecies}}{integer vector of spp IDs}

\item{\code{indTrait}}{numeric matrix of trait values for each individual (= rows)}

\item{\code{indSeqs}}{character vector with sequences for each individual}
}}

