% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaComm.R
\docType{class}
\name{metaComm-class}
\alias{metaComm-class}
\alias{metaComm}
\title{Meta community}
\usage{
metaComm(sppAbund, sppTrait)
}
\description{
An S4 class to specify the state of a meta community
}
\section{Slots}{

\describe{
\item{\code{sppAbund}}{numeric vector with abundances for each spp}

\item{\code{sppTrait}}{numeric matrix with traits for each spp}
}}

