% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleFoo.R
\docType{class}
\name{roleFoo-class}
\alias{roleFoo-class}
\title{A set of objects representing the state of a roleModel}
\description{
Contains community states, phylo state, and params
}
\section{Slots}{

\describe{
\item{\code{localComm}}{an object of class \code{localComm}}

\item{\code{metaComm}}{an object of class \code{metaComm}}

\item{\code{phylo}}{an object of class \code{rolePhylo}}

\item{\code{params}}{an object of class \code{roleParams}}
}}

