% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolePhylo.R
\name{rolePhylo}
\alias{rolePhylo}
\title{Specify a RoLE model phylogeny}
\usage{
rolePhylo(n, e, l, alive, tipNames)
}
\arguments{
\item{n}{number of tips}

\item{e}{edge matrix; two columns give ancestor, child pair}

\item{l}{numeric vector of edge lengths}

\item{alive}{logical vector indicating whether tips are extant or not}

\item{tipNames}{character vector of tip names}
}
\description{
Specify a RoLE model phylogeny
}
